Separate nested class mixins
