+++ common/src/main/java/org/mcaccess/minecraftaccess/compat/mixin/clothconfig/ClothConfigScreenMixin.java
@@ -123,38 +123,4 @@ abstract class ClothConfigScreenMixin extends AbstractTabbedConfigScreen {
             tabButton.onPress(new MouseButtonInfo(0, 0));
         }
     }
-
-    @SuppressWarnings("rawtypes")
-    @Mixin(value = ClothConfigScreen.ListWidget.class, remap = false)
-    abstract static class ListWidgetMixin extends DynamicElementListWidget {
-        @Shadow
-        @Final
-        private AbstractConfigScreen screen;
-
-        protected ListWidgetMixin(Minecraft client, int width, int height, int top, int bottom, ResourceLocation backgroundLocation) {
-            super(client, width, height, top, bottom, backgroundLocation);
-        }
-
-        @Override
-        @Shadow
-        public abstract @NotNull List<GuiEventListener> children();
-
-        /**
-         * Focus on the last option if navigating backward from cancel button
-         */
-        @Override
-        public ComponentPath nextFocusPath(FocusNavigationEvent event) {
-            List<GuiEventListener> children = children();
-            if (!isFocused() && !children.isEmpty() && NavigationUtils.isDirectionBackward(event)) {
-                setFocused(null);
-                return ComponentPath.path(this, NavigationUtils.getFocusPathStartFrom(children.getLast(), event));
-            }
-            return super.nextFocusPath(event);
-        }
-
-        @Override
-        public boolean isFocused() {
-            return screen.getFocused() == this;
-        }
-    }
 }
