+++ common/src/main/resources/minecraft_access-compat.mixins.json
@@ -5,15 +5,15 @@
     "compatibilityLevel": "JAVA_21",
     "client": [
         "clothconfig.AbstractConfigListEntryMixin",
+        "clothconfig.CategoryLabelWidgetMixin",
         "clothconfig.CleanMouseClickingFocus",
         "clothconfig.ClothConfigScreenAccessor",
         "clothconfig.ClothConfigScreenMixin",
-        "clothconfig.ClothConfigScreenMixin$ListWidgetMixin",
         "clothconfig.ClothConfigTabButtonMixin",
         "clothconfig.ContainerEventHandlerMixin",
+        "clothconfig.ListWidgetMixin",
         "clothconfig.ScreenNarrationCollectorOutputMixin",
-        "clothconfig.SubCategoryListEntryMixin",
-        "clothconfig.SubCategoryListEntryMixin$CategoryLabelWidgetMixin"
+        "clothconfig.SubCategoryListEntryMixin"
     ],
     "injectors": {
         "defaultRequire": 1
