+++ common/src/main/resources/minecraft_access.mixins.json
@@ -22,7 +22,6 @@
         "CheckboxMixin",
         "ClientPacketListenerMixin",
         "ClientTextTooltipAccessor",
-        "CommandSuggestionsSuggestionsListMixin",
         "CreativeModeInventoryScreenAccessor",
         "EditBoxAccessor",
         "EditBoxMixin",
@@ -61,6 +60,7 @@
         "SingleItemRecipeAccessor",
         "SlotAccessor",
         "StonecutterScreenAccessor",
+        "SuggestionsListMixin",
         "SystemToastAccessor",
         "TamableAnimalMixin",
         "TextFieldHelperMixin",
